#16 Change int ->long

+++ code/basic_functions.c
@@ -65,7 +65,7 @@ void set_cuda(int debug_flag){
     printf("Active card:\t%s\n", &properties.name);
 }
 
-double linear_interpolate(double *x_vector, int x_length, double *y_vector, int y_length, double x_value){
+double linear_interpolate(double *x_vector, long x_length, double *y_vector, long y_length, double x_value){
     int i;
     if (x_length != y_length)   printf("ERROR: in interpolation. Two input vectors have different length.");
     for (i=1; (i<x_length) && (x_vector[i-1]>x_value); ++i);
