+++ code/dataio/parameter_reader.c
@@ -39,7 +39,7 @@ void init_taiga_props(char* par_name, char* par_value, BeamProp *beam, ShotProp 
     else if (!strcmp(par_name, "step_device"))              run->step_device = par_value_d;
     else if (!strcmp(par_name, "step_host"))                run->step_host = par_value_d;
     else if (!strcmp(par_name, "particles")){
-        run->particle_number = par_value_d;
+        run->particle_number = (long)par_value_d;
         run->block_number    = par_value_d/run->block_size+1;
     }
 }
